core/vm: move auth* ops back to eips
